search2/requirements.txt
faiss-cpu==1.7.4 
sentence-transformers==2.2.2 
rank-bm25==0.2.2 
faker==13.15.0 
nltk==3.8.1 
numpy==1.24.3